[fix] move themedata_manager.dart on constants
