+++ apps/mobiles/whilabel/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:whilabel/presentation/resources/routes_manager.dart';
-import 'package:whilabel/presentation/resources/themedata_manager.dart';
-
+import 'package:whilabel/screens/constants/routes_manager.dart';
+import 'package:whilabel/screens/constants/themedata_manager.dart';
 
 void main() {
   runApp(const MainApp());
